+++ Changelog.md
@@ -1,3 +1,3 @@
-# gov-ua-ui changelog from v0.5.0. to v0.5.1.
+# gov-ua-ui changelog from v0.5.1. to v0.6.0.
 
-1. Added monthes select to DatePicker
+1. Added new Pagination element